+++ test/server_test.dart
@@ -3,20 +3,20 @@ import 'package:les/les.dart';
 void main() {
   Server()
   ..use(bodyParser)
-  ..use(Middleware((ctx) {
-    print(ctx.reqBody);
-  }))
-    //..add(routes)
+  ..use(buildStaticFilesHandler())
+    ..add(routes)
     ..listen(7575);
 }
 
 final routes = Router('/')
   ..add(Route(
         r':a(\d+)',
-        (ctx) => ctx.send('Hello get')
+        (ctx) async {
+          await ctx.send(ctx.files['static/image.png'].readAsBytesSync());
+        }
       ))
   ..add(Route(
-      'af/af',
+      'af',
       (ctx) => ctx.send('Hello post'),
       method: HttpMethod.delete
       ));